DeleteConfirmed now savesChanges, updating the database

+++ CPW219-AspnetMVC-CRUD-Debugging/Controllers/ProductsController.cs
@@ -76,8 +76,14 @@ namespace CPW219_AspnetMVC_CRUD_Debugging.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var product = await _context.Product.FindAsync(id);
-            _context.Product.Remove(product);
-            return RedirectToAction(nameof(Index));
+            if (product != null)
+            {
+                _context.Product.Remove(product);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return RedirectToAction("Index");
+
         }
 
         private bool ProductExists(int id)
